@@ -34,7 +34,7 @@
     "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-image-video-lightbox": "^3.0.1"
+    "yet-another-react-lightbox": "^3.15.6"
   },
   "devDependencies": {
     "@docusaurus/eslint-plugin": "^2.4.0",
@@ -63,6 +63,6 @@
     ]
   },
   "engines": {
-    "node": ">=16.14"
+    "node": ">=18.0"
   }
 }
